Make Airflow CTL tests more like pytests
