+++ airflow-ctl-tests/tests/airflowctl_tests/conftest.py
@@ -20,7 +20,6 @@ import os
 import subprocess
 import sys
 
-import pytest
 from python_on_whales import DockerClient, docker
 
 from airflowctl_tests import console
@@ -201,112 +200,3 @@ def pytest_sessionfinish(session, exitstatus):
     """Tear down test environment at the end of the pytest session."""
     if not os.environ.get("SKIP_DOCKER_COMPOSE_DELETION"):
         docker_compose_down()
-
-
-# Fixtures for tests
-@pytest.fixture
-def login_command():
-    # Passing password via command line is insecure but acceptable for testing purposes
-    # Please do not do this in production, it enables possibility of exposing your credentials
-    return "auth login --username airflow --password airflow"
-
-
-@pytest.fixture
-def login_output():
-    return "Login successful! Welcome to airflowctl!"
-
-
-@pytest.fixture
-def date_param():
-    import random
-    from datetime import datetime, timedelta
-
-    from dateutil.relativedelta import relativedelta
-
-    # original datetime string
-    dt_str = "2025-10-25T00:02:00+00:00"
-
-    # parse to datetime object
-    dt = datetime.fromisoformat(dt_str)
-
-    # boundaries
-    start = dt - relativedelta(months=1)
-    end = dt + relativedelta(months=1)
-
-    # pick random time between start and end
-    delta = end - start
-    random_seconds = random.randint(0, int(delta.total_seconds()))
-    random_dt = start + timedelta(seconds=random_seconds)
-    return random_dt.isoformat()
-
-
-@pytest.fixture
-def test_commands(login_command, date_param):
-    # Define test commands to run with actual running API server
-    return [
-        login_command,
-        # Assets commands
-        "assets list",
-        "assets get --asset-id=1",
-        "assets create-event --asset-id=1",
-        # Backfill commands
-        "backfill list",
-        # Config commands
-        "config get --section core --option executor",
-        "config list",
-        "config lint",
-        # Connections commands
-        "connections create --connection-id=test_con --conn-type=mysql --password=TEST_PASS -o json",
-        "connections list",
-        "connections list -o yaml",
-        "connections list -o table",
-        "connections get --conn-id=test_con",
-        "connections get --conn-id=test_con -o json",
-        "connections update --connection-id=test_con --conn-type=postgres",
-        "connections import tests/airflowctl_tests/fixtures/test_connections.json",
-        "connections delete --conn-id=test_con",
-        "connections delete --conn-id=test_import_conn",
-        # DAGs commands
-        "dags list",
-        "dags get --dag-id=example_bash_operator",
-        "dags get-details --dag-id=example_bash_operator",
-        "dags get-stats --dag-ids=example_bash_operator",
-        "dags get-version --dag-id=example_bash_operator --version-number=1",
-        "dags list-import-errors",
-        "dags list-version --dag-id=example_bash_operator",
-        "dags list-warning",
-        # Order of trigger and pause/unpause is important for test stability because state checked
-        f"dags trigger --dag-id=example_bash_operator --logical-date={date_param} --run-after={date_param}",
-        "dags pause example_bash_operator",
-        "dags unpause example_bash_operator",
-        # DAG Run commands
-        f'dagrun get --dag-id=example_bash_operator --dag-run-id="manual__{date_param}"',
-        "dags update --dag-id=example_bash_operator --no-is-paused",
-        # DAG Run commands
-        "dagrun list --dag-id example_bash_operator --state success --limit=1",
-        # Jobs commands
-        "jobs list",
-        # Pools commands
-        "pools create --name=test_pool --slots=5",
-        "pools list",
-        "pools get --pool-name=test_pool",
-        "pools get --pool-name=test_pool -o yaml",
-        "pools update --pool=test_pool --slots=10",
-        "pools import tests/airflowctl_tests/fixtures/test_pools.json",
-        "pools delete --pool=test_pool",
-        "pools delete --pool=test_import_pool",
-        # Providers commands
-        "providers list",
-        # Variables commands
-        "variables create --key=test_key --value=test_value",
-        "variables list",
-        "variables get --variable-key=test_key",
-        "variables get --variable-key=test_key -o table",
-        "variables update --key=test_key --value=updated_value",
-        "variables import tests/airflowctl_tests/fixtures/test_variables.json",
-        "variables delete --variable-key=test_key",
-        "variables delete --variable-key=test_import_var",
-        "variables delete --variable-key=test_import_var_with_desc",
-        # Version command
-        "version --remote",
-    ]
